Cast the ioctl define to the type of the variable being switched on.
git-svn-id: svn+ssh://svn.freebsd.org/base/head@170490 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/net/if_spppsubr.c
@@ -5176,7 +5176,7 @@ sppp_params(struct sppp *sp, u_long cmd, void *data)
 	}
 
 	switch (subcmd) {
-	case (int)SPPPIOGDEFS:
+	case (u_long)SPPPIOGDEFS:
 		if (cmd != SIOCGIFGENERIC) {
 			rv = EINVAL;
 			break;
@@ -5211,7 +5211,7 @@ sppp_params(struct sppp *sp, u_long cmd, void *data)
 			     sizeof(struct spppreq));
 		break;
 
-	case (int)SPPPIOSDEFS:
+	case (u_long)SPPPIOSDEFS:
 		if (cmd != SIOCSIFGENERIC) {
 			rv = EINVAL;
 			break;
